backend/requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1

# Database
psycopg2[binary]

# Production (opcional)
gunicorn==22.0.0
python-dotenv==1.0.1
